+++ com/as.infrastructure/include/CanIf_Cbk.h
@@ -24,7 +24,7 @@
 #define CANIF_CBK_H_
 
 void CanIf_TxConfirmation( PduIdType canTxPduId );
-void CanIf_RxIndication( uint8 Hrh, Can_IdType CanId, uint8 CanDlc, const uint8 *CanSduPtr );
+void CanIf_RxIndication( uint16 Hrh, Can_IdType CanId, uint8 CanDlc, const uint8 *CanSduPtr );
 
 #if ( CANIF_TRANSMIT_CANCELLATION == STD_ON )
 void CanIf_CancelTxConfirmation( const Can_PduType *PduInfoPtr );
